+++ CHANGELOG.md
@@ -6,8 +6,16 @@ New developments that use __graphql goal/task__ should use the parameters below,
 * separateUtilityClasses: true
 
 
+# Not released yet 
+
 Both mode (client and server):
-* Attributes of input types that are list, boolean or custom scalars where not properly managed
+* Upgrade of graphql-java from v14.0 to v15.0
+* Main improvement for this: the plugin now accepts that an interface implements one or more interfaces
+* Attributes of input types that are enum are now properly managed
+
+Server mode:
+* The generated code would not compile if a GraphQL interface is defined, but not used, in the given GraphQL schema
+
 
 # 1.9